requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==3.2.4
django-cors-headers==3.7.0
django-cron==0.6.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
idna==3.10
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
gunicorn==23.0.0
uvicorn==0.32.1
twilio===9.4.1